requirements.txt
streamlit
ultralytics
pytesseract
opencv-python
pillow
numpy